(GH-24) Support large file uploads
If the file is above the public api size limit and the user has
selected `RestrictSizeLimits = false`, then attempt to use the upload
url. If the user does not have access, they will receive the regular
access denied message:  "AccessDeniedException: You don't have access to
the service. Make sure your API key is working correctly."

If the user does have access, it will allow the large file to be
uploaded using the private API url. Since the base url is
different, override the RestClient with a large upload RestClient.